[Storage]Update ServerCommon library to v12 storage SDKs (#430)
* Start update to new SDK

* Move most of the storage to new SDK.

* migrate transfer management in AzureStorage

* Update Queue Implmentation to v12 SDK

* Migrate to v12 SDK

* Migrate Tables to new SDK

* Switch to generic Blob isntead of BlockBlob

* Temporarily disable emu tests and add new interface ones.

* Adding Save tests.

* Point Transfermanager to temp directory.

* Potential fix for incorrect container Uri.

* Added some explciit typing.

* Add lease adapters to storage SDK types

* Update List to use metadata bool. Clean up some comments

* Update language version. Add async list

* Update StorageExtension to new c# syntax.

* Remove unneeded underlying TableEntity.

* remove some unclear var usage.

* Start update to new SDK

* Move most of the storage to new SDK.

* migrate transfer management in AzureStorage

* Update Queue Implmentation to v12 SDK

* Migrate to v12 SDK

* Migrate Tables to new SDK

* Switch to generic Blob isntead of BlockBlob

* Temporarily disable emu tests and add new interface ones.

* Adding Save tests.

* Point Transfermanager to temp directory.

* Potential fix for incorrect container Uri.

* Added some explciit typing.

* Add lease adapters to storage SDK types

* Update List to use metadata bool. Clean up some comments

* Update language version. Add async list

* Update StorageExtension to new c# syntax.

* Remove unneeded underlying TableEntity.

* remove some unclear var usage.

* BlobClient -> blockBlobClient for perf.

* Cleaning up some duplicated messages. Fix setting headers AFTER upload.

* Trivially implement ListAsync for other Storages.

* Update tests for new changes.

* Clean up old tests.

---------

+++ src/NuGet.Services.Storage/StorageQueueMessage.cs
@@ -1,14 +1,16 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
+
 namespace NuGet.Services.Storage
 {
     public class StorageQueueMessage
     {
         public string Contents { get; }
-        public int DequeueCount { get; }
+        public long DequeueCount { get; }
 
-        public StorageQueueMessage(string contents, int dequeueCount)
+        public StorageQueueMessage(string contents, long dequeueCount)
         {
             Contents = contents;
             DequeueCount = dequeueCount;
@@ -18,9 +20,9 @@ namespace NuGet.Services.Storage
     public class StorageQueueMessage<T>
     {
         public T Contents { get; }
-        public int DequeueCount { get; }
+        public long DequeueCount { get; }
 
-        public StorageQueueMessage(T contents, int dequeueCount)
+        public StorageQueueMessage(T contents, long dequeueCount)
         {
             Contents = contents;
             DequeueCount = dequeueCount;
